+++ tests/src/legacy/cfg-attr/lib.rs
@@ -57,3 +57,21 @@ impl Double for Counter {
         x + x
     }
 }
+
+#[cfg_attr(hax, hax_lib::attributes)]
+struct Refined<const LEN: usize> {
+    #[cfg_attr(hax, hax_lib::refine(hax_lib::forall(|i: usize| hax_lib::implies(
+        i < indices.len(),
+        (indices[i] as usize) < 2
+    ))))]
+    indices: [u8; LEN],
+    #[cfg_attr(hax_backend_fstar, hax_lib::refine(x < 5))]
+    x: u8,
+}
+
+#[cfg_attr(hax, hax_lib::attributes)]
+struct Reordered {
+    x: u8,
+    #[cfg_attr(hax, hax_lib::order(-1))]
+    y: u8,
+}