+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+Changes to the hax-lib crate:
+ - Support `refine` and `order` written behind a `cfg_attr` on the fields of an item annotated
+   with `#[hax_lib::attributes]`, keeping the `cfg_attr` predicate (#1496)
+ - Strip `order` from enum variant fields in builds without `--cfg hax`
+
 ## [0.4.1] - 2026-09-23
 
 ### Fixed